实现了admin端和app端 用户相关功能（包含注册，登录，推出。涉及token令牌过期黑名单，鉴权，），角色相关功能

+++ web/web-app/src/main/java/com/huang/web/app/vo/profile/ProfileDetailVO.java
@@ -55,10 +55,10 @@ public class ProfileDetailVO {
     private String occupation;
 
     @Schema(description = "身高（厘米）")
-    private Integer height;
+    private Double height;
 
     @Schema(description = "体重（公斤）")
-    private Integer weight;
+    private Double weight;
 
     @Schema(description = "状态：0-禁用，1-正常")
     private Integer status;
